Fix expression arena inflation, imPol memoization, and opening point propagation
Three interconnected bugs caused the first AIR (Dma) output to differ
from the golden reference:

1. Expression arena inflation: pilout parsing and constraint/FRI/imPol
   builders pushed every helper node (challenge, exp refs, zi, cm, etc.)
   as separate arena entries, but JS keeps them as inline child objects.
   Introduced ExprChild enum (Id | Inline) so children can be stored
   inline without inflating the arena. Updated pilout_info, constraint_poly,
   im_polynomials, and fri_poly to use inline children, matching JS
   semantics where only composite expressions are pushed to the array.

2. Intermediate polynomial memoization: the memo key now includes the
   expression ID (matching JS per-expression cache pattern
   exp.res[absoluteMax][JSON.stringify(imPols)]). Also fixed qDeg output
   serialization which was incorrectly using base-field column count
   instead of the polynomial degree from the imPol search.

3. Opening point -1 propagation: added info_computed flag to Expression
   to replace the fragile heuristic guard in add_info_expressions. The
   old guard incorrectly skipped exp nodes whose stage was pre-set to 1
   by prepare_pil, preventing rows_offsets (including -1) from
   propagating. Also changed opening field type from usize to i64 to
   support negative opening points in xDivXSubXi nodes.

+++ pil2-stark-setup/src/types.rs
@@ -168,7 +168,7 @@ pub struct CodeRef {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub commit_id: Option<usize>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub opening: Option<usize>,
+    pub opening: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub boundary_id: Option<usize>,
     #[serde(skip_serializing_if = "Option::is_none")]
